Adopt ruff 0.16 and modernize the codebase to its default rules (#345)
ruff 0.16 broadens the default rule set (pyupgrade, isort, comprehensions and
more). Bump the dev pin to ruff>=0.16.0 and apply the resulting fixes across
the tree: PEP 585 built-in generic annotations, unquoted annotations under
from-future annotations, sorted imports and the assorted small lint fixes. A
re-export shim that relied on redundant import aliases now declares __all__
explicitly. The api reference is regenerated for the modernized signatures.
No runtime behavior changes; the full test suite passes.

+++ tests/broadcast/test_broadcast_plot_i18n.py
@@ -8,17 +8,17 @@ import matplotlib
 
 matplotlib.use("Agg")
 
-import matplotlib.pyplot as plt  # noqa: E402
-import numpy as np  # noqa: E402
-import pytest  # noqa: E402
+import matplotlib.pyplot as plt
+import numpy as np
+import pytest
 
-from phonometry.broadcast import program_loudness  # noqa: E402
+from phonometry.broadcast import program_loudness
 
 FS = 48000
 
 
 def _sine(level_dbfs: float, duration: float) -> np.ndarray:
-    t = np.arange(int(round(duration * FS))) / FS
+    t = np.arange(round(duration * FS)) / FS
     return 10.0 ** (level_dbfs / 20.0) * np.sin(2.0 * np.pi * 1000.0 * t)
